Use dfs taskgraph get order
* taskgraph get order:
  * use dfs instead of bfs
* mg make:
  * use tensors on memory for deps in register usage
* llama model parallel: use f16
* gpu llama: added diff partitioning strategies

+++ llama/gpu_model_parallel.cc
@@ -308,7 +308,7 @@ void main_rank_zero(
   int ngpus);
 
 int main(int argc, char** argv) {
-  set_default_dtype(dtype_t::f32);
+  set_default_dtype(dtype_t::f16);
   args_t pargs(argc, argv);
 
   string which_model = pargs.get<string>("model");
